Attempt to switch SD cards to high-speed (50MHz) mode
... and decrease prescaler

This assumes that the _4 prescaler is a clock speed > 25MHz (the default
speed). My card that would previously fail with the faster setting still
works with this change as it returns failure from the change command.
This doesn't check if the card supports spec version 1.10 which is
required for this command.